Add `py` package as explicit dev dependency
It appears that this was previously a transitive dependency of
one of the other packages installed in the tox environment, but is
no longer (most likely due to the package being deprecated). This
plugin only depends on the package for mypy typing and for backward
compatibility with versions of PyTest prior to 7.0.