Avoid fetching manifest in xdist workers on failure
xdist workers don't have access to the API key and cannot fetch the
manifest. Prior to this change, the xdist worker would attempt to
fetch the manifest if the fetch attempt in the parent process failed.
This change fixes the bug so that xdist workers will never attempt
to fetch the manifest, even if the fetch in the parent process failed.