Support test run payloads over 6MB
AWS Lambda limits payloads to 6MB, which puts an upper bound
on the size of the Unflakable backend API's request bodies. This
change leverages a newer backend API endpoint to upload test runs
to presigned S3 URLs, which are not subject to AWS Lambda limits.
The backend then retrieves the upload directly from S3.